Remote storage first pass
Can now create a new profile from a set of images and download the
images for profiles filtered by id (so you don't have to re-download
one's you already have locally).
When you download a profile's images, the face identification index file
is automatically updated (or created if missing).
Note that config.js is currently set up for local testing and the DB_URL
will need to be changed for deployment.